migrated all testing logic out of the Client
